% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashdress.R
\name{add_parcel_id}
\alias{add_parcel_id}
\title{Add CAGIS Parcel ID using hashdress method}
\usage{
add_parcel_id(.x, quiet = TRUE)
}
\arguments{
\item{.x}{a vector of address character strings}

\item{quiet}{logical; suppress intermediate DeGAUSS console output?}
}
\value{
a list the same length as .x; each item will be a vector of \code{parcel_id}s
because more than one parcel can be matched to a given address
}
\description{
A vector of address character strings are expanded and hashed
into a set of hashdresses to match to an internal lookup
table containing hashdresses for CAGIS Parcels.
}
\details{
Addresses will be parsed and only the \code{street_number} and \code{street_name}
\emph{from addresses with Hamilton County ZIP codes} are used to
expand, hashdress, and match to a parcel identifier. Addresses not in Hamilton
County will return a missing parcel identifier without an attempted match.
One address can be linked to more than one parcel (e.g.,
"323 Fifth" on https://wedge3.hcauditor.org/search_results).
}
\examples{
\dontrun{
d <- data.frame(address = c(
  "3937 Rose Hill Ave Cincinnati OH 45229",
  "424 Klotter Ave Cincinnati OH 45214",
  "3328 Bauerwoods Dr Cincinnati OH 45251"
))
add_parcel_id(d, quiet = TRUE)
}
}
