% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashdress.R
\name{address_expand}
\alias{address_expand}
\title{'expand' addresses containing abbreviations into all possible addresses}
\usage{
address_expand(.x, hashdress = TRUE, quiet = TRUE)
}
\arguments{
\item{.x}{a tibble containing an \code{address} column}

\item{hashdress}{logical; also compute hashdress?}

\item{quiet}{logical; suppress intermediate DeGAUSS console output?}
}
\description{
The DeGAUSS \href{https://degauss.org/postal/}{\code{postal}} container is used first
to create clean addresses consisting of the parsed \code{house_number}, \code{road},
and first five digits of \code{postcode}. It is used again to expand these based on abbreviations.
Because each input address will likely result in more than one expanded address,
the newly added \code{expanded_addresses} column is a list-col.
}
\details{
By default, each \code{expanded_address} is hashed using the 'spookyhash' algorithm and also
returned as a list col. These combinations of hashes for all expanded address
(i.e. "hashdress") can be used to link to other hashdress data resources (see \code{add_parcel_id()})

Each call to DeGAUSS is cached to disk (\code{data-raw} folder in working directory),
making repetative function calls on the same data nearly instant.
}
\examples{
d <-
  tibble::tibble(address = c(
    "224 Woolper Ave Cincinnati OH 45220",
    "222 East Central Parkway Cincinnati OH 45220",
    "352 Helen St Cincinnati OH 45202",
    "5377 Bahama Ter Apt 1 Cincinnati Ohio 45223",
    "5377 Bahama Te Apt 1 Cincinnati Ohio 45223",
    "1851 Campbell Dr Hamilton Ohio 45011",
    "2 Maplewood Dr Ryland Heights, KY 41015"
  ))
address_expand(d)
set cache for degauss_run
}
