% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{build}
\alias{build}
\alias{parcel_build}
\alias{parcel_dev}
\title{Run Srcipt}
\usage{
parcel_build()

parcel_dev()
}
\description{
Run \code{package.json} scripts.
}
\section{Functions}{
\itemize{
\item \code{parcel_dev}: Run the \code{dev} server
}}

