% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cagis_parcels}
\alias{cagis_parcels}
\title{CAGIS parcels}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 286059 rows and 13 columns.
}
\source{
Cincinnati Area Geographic Information System (CAGIS: https://cagismaps.hamilton-co.org/cagisportal/mapdata/download)
}
\usage{
cagis_parcels
}
\description{
A curated property-level data resource derived from the Hamilton County, OH Auditor data distributed through CAGIS Open Data. See \code{data-raw/tabular-data-resource.yaml} for detailed metadata.
}
\keyword{datasets}
