% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{check_installed}
\alias{check_installed}
\title{Check whether a required package is installed}
\usage{
check_installed(pkg = c("npm", "yarn"))
}
\description{
Check whether a required package is installed
}
\keyword{internal}
