% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine}
\alias{engine}
\alias{set_engine}
\alias{get_engine}
\title{Set Engine}
\usage{
set_engine(engine = c("npm", "yarn"))

get_engine()
}
\arguments{
\item{engine}{Engine to use, either \code{npm}, or \code{yarn}.}
}
\description{
Define whether you want to use npm or yarn.
The default engine is \code{npm}, it can be changed
globally by setting the environment variable
\code{PARCEL_ENGINE} in \code{.Rprofile}.
}
\section{Functions}{
\itemize{
\item \code{get_engine}: Retrieve defined engine.
}}

