% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{expand_addresses}
\alias{expand_addresses}
\title{expand addresses using postal}
\usage{
expand_addresses(addresses, quiet = TRUE)
}
\arguments{
\item{addresses}{a vector of address strings}

\item{quiet}{suppress DeGAUSS output inside of R?}
}
\value{
a list the same length as the input \code{addresses} where each
element is a vector of expanded addresses
}
\description{
The DeGAUSS \href{https://degauss.org/postal/}{\code{postal}} container is used
to expand a vector of addresses based on abbreviations.
Each call to DeGAUSS is cached to disk (\code{data-raw} folder in working directory),
making repetative function calls on the same data nearly instant.
}
