% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_parcel.R
\name{get_parcel_data}
\alias{get_parcel_data}
\title{return parcel data for input addresses}
\usage{
get_parcel_data(x)
}
\arguments{
\item{x}{a vector of address character strings}
}
\value{
a tibble with the \code{input_address}es defined in \code{x} in the first column,
and columns corresponding to matched parcel characteristics
}
\description{
This helper function produces a tibble of parcel data for an input vector of addresses.
The \code{link_parcel()} function returns all possible matches above the \code{threshold} for each
input address and this function chooses the single best match based on the maximum score.
Note that one address can be linked to more than one parcel with the same match score (e.g.,
"323 Fifth" on https://wedge3.hcauditor.org/search_results). In this case,
a special identifier, \code{TIED_MATCHES} is returned instead of a missing \code{parcel_id}.
For finer control of selecting matched parcels based on scores, use \code{link_parcel()}
}
