% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashdress.R
\name{hashdress}
\alias{hashdress}
\title{hash addresses}
\usage{
hashdress(.x)
}
\arguments{
\item{.x}{a list of vectors of address character strings}
}
\value{
a list the same length as .x
}
\description{
These combinations of hashes for all expanded address
(i.e. "hashdress") can be used to link to other addresses hashdressed
using the same version of this package.
}
