% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_parcel.R
\name{link_parcel}
\alias{link_parcel}
\title{link addresses to CAGIS parcel identifiers}
\usage{
link_parcel(x)
}
\arguments{
\item{x}{a vector of address character strings}
}
\value{
a tibble with a column of \emph{unique}, matched addresses input as \code{x} along
with columns for their \code{parcel_id}(s) and matching \code{score}(s) (use this as a lookup
table for assigning parcel_id in other workflows, making decisions about what to do
with multiple matches and matching thresholds, etc.)
}
\description{
This function uses the trained dedupe model included with the package
to link one or more parcel identifiers to a vector of input
addresses.
}
\details{
Note that one address can be linked to more than one parcel (e.g.,
"323 Fifth" on https://wedge3.hcauditor.org/search_results). In this case,
the input address will have multiple rows, one for each of the multiple matches.
}
