% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_address.R
\name{tag_address}
\alias{tag_address}
\title{tag components of an address}
\usage{
tag_address(address, clean = TRUE)
}
\arguments{
\item{address}{a character string that is a United States mailing address}

\item{clean}{clean addresses with \code{clean_address()} prior to tagging?}
}
\value{
a tibble with \code{street_number}, \code{street_name}, \code{city}, \code{state}, and \code{zip_code} columns
}
\description{
This function relies on \code{usaddress} python library https://usaddress.readthedocs.io/en/latest/
It can be installed to a python virtual environment specific to R with:
\code{py_install("usaddress", pip = TRUE)} (See the README for more details on installing
and managing non-system installations of python with {reticulate}.
}
\details{
This function uses a custom tag mapping to combine address components into the columns in the returned tibble
(see https://usaddress.readthedocs.io/en/latest/#details for full definition of components):
\itemize{
\item \code{street_number}: \code{AddressNumber}, \code{AddressNumberPrefix}, \code{AddressNumberSuffix}
\item \code{street_name}: \code{StreetName}, \code{StreetNamePreDirectional}, \code{StreetNamePostDirectional}, \code{StreetNamePostModifier}, \code{StreetNamePostType}
\item \code{city}: \code{PlaceName}
\item \code{state}: \code{StateName}
\item \code{zip}: the \strong{first five characters} of \code{ZipCode}
}

If an address is not classified as a \verb{Street Address} (i.e. \code{Intersection}, \verb{PO Box}, or \code{Ambiguous}),
then the columns in the returned component tibble will all be missing.
}
