% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{type2output}
\alias{type2output}
\title{Get Output Path from dir}
\usage{
type2output(type)
}
\arguments{
\item{type}{Project type as returned by \code{\link[=get_project_type]{get_project_type()}}.}
}
\description{
Get Output Path from dir
}
\keyword{internal}
