% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashdress.R
\name{add_parcel_id}
\alias{add_parcel_id}
\title{Add CAGIS Parcel ID}
\usage{
add_parcel_id(.x, quiet = TRUE)
}
\arguments{
\item{.x}{tibble/data.frame with a column containing addresses, called "address"}

\item{quiet}{logical; suppress intermediate DeGAUSS console output?}
}
\value{
.x with additional parcel_id column; this will be a list-col because more than
one parcel can be matched to a given address
}
\description{
The hashdress() function is used to add a set of hashdresses for
the parsed house_number and road from each input address in \code{.x$address}.
The hashdresses are compared to the
set of hashdresses in \code{cagis_hashdresses} for matches.
}
\details{
One address can be linked to more than one parcel (e.g.,
"323 Fifth" on https://wedge3.hcauditor.org/search_results)
}
\examples{
d <- data.frame(address = c(
  "3937 Rose Hill Ave Cincinnati OH 45229",
  "424 Klotter Ave Cincinnati OH 45214",
  "3328 Bauerwoods Dr Cincinnati OH 45251"
))
add_parcel_id(d, quiet = TRUE)
}
