% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cagis_hashdresses}
\alias{cagis_hashdresses}
\title{CAGIS hashdresses}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 528483 rows and 4 columns.
}
\usage{
cagis_hashdresses
}
\description{
The parcel identifiers and a property address (consisting of
property_addr_number, property_addr_street, property_addr_suffix)
for each parcel are \code{hashdress()}ed using the parsed street number and street name.
This object is used to match parcel identifiers
to hashdresses computed on other, real-world addresses. Note that
the five digit ZIP code is not included in CAGIS data, and wasn't used to
compute the hashdress.  These hashdresses are specific to Hamilton
County, OH.
}
\keyword{datasets}
