% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_address.R
\name{clean_address}
\alias{clean_address}
\title{clean address text}
\usage{
clean_address(.x)
}
\arguments{
\item{.x}{a vector of address character strings}
}
\value{
a vector of cleaned addresses
}
\description{
convert to lowercase, remove non-alphanumeric characters and excess whitespace
(adapted from degauss-org/dht::clean_address)
}
