% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_stub.R
\name{create_address_stub}
\alias{create_address_stub}
\title{prepare addresses for matching to CAGIS Parcel data}
\usage{
create_address_stub(.x, filter_zip = TRUE)
}
\arguments{
\item{.x}{a vector of address character strings}

\item{filter_zip}{force addresses with non-Hamilton ZIP codes to have a missing address_stub?}
}
\value{
a vector of cleaned address stubs (street_number + street_name)
}
\description{
Input addresses are cleaned (convert to lowercase, remove
non-alphanumeric characters and excess whitespace) and then
tagged into components. The \code{street_number} and \code{street_name}
components are pasted together to create the address stub.
If either the street_number or street_name are missing
then the address_stub will be returned as missing.
If \code{filter_zip} is TRUE, then addresses without a parsed
5-digit ZIP code in Hamilton County will have a missing address stub.
}
