% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_apt.R
\name{link_apt}
\alias{link_apt}
\title{Link one address to parcel pseudo-identifiers for apartment complexes}
\usage{
link_apt(x)
}
\arguments{
\item{x}{a single address character string}
}
\value{
apt pseudo-identifier character string; \code{NA} if not matched
}
\description{
To match a parcel to an apartment complex pseudo-identifier, it must contain:
\itemize{
\item a Hamilton County ZIP code
\item a street name matching the street names in \code{parcel:::apt_defs}
\item a street number within the ranges for each pseudo-identifier in \code{parcel:::apt_defs}
}
}
