% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_apt.R
\name{link_apt}
\alias{link_apt}
\title{Link one addresses to custom parcel identifiers based on street names and ranges}
\usage{
link_apt(x)
}
\arguments{
\item{x}{a vector of address character strings}
}
\description{
This function will probably take a long time to map over large numbers of addresses.
Reduce the use of this function by only applying it to addresses that are unmatched
with \code{link_parcel()}, but have a Cincinnati ZIP code. (
}
