% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaddress.R
\name{tag_address}
\alias{tag_address}
\title{tag components of an address using the \code{usaddress} python library}
\usage{
tag_address(address)
}
\arguments{
\item{address}{a character string that is a United States mailing address}
}
\value{
a tibble with \code{street_number}, \code{street_name}, \code{city}, \code{state}, and \code{zip_code} columns
}
\description{
This function relies on usaddress package https://usaddress.readthedocs.io/en/latest/
It can be installed to a python virtual environment specific to R with:
\code{py_install("usaddress", pip = TRUE)}
This function uses a custom tag mapping to categorize (and combine some) address
components (see https://usaddress.readthedocs.io/en/latest/#details).
If an address is not classified as a \verb{Street Address} (i.e. \code{Intersection}, \verb{PO Box}, or \code{Ambiguous}),
then the columns in the returned component tibble will all be NA
}
