% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaddress.R
\name{tag_address}
\alias{tag_address}
\title{tag components of an address using the \code{usaddress} python library}
\usage{
tag_address(address)
}
\arguments{
\item{address}{a character string that is a United States mailing address}
}
\value{
a tibble with \code{street_number}, \code{street_name}, \code{city}, \code{state}, and \code{zip_code} columns
}
\description{
This function relies on usaddress package https://usaddress.readthedocs.io/en/latest/
It can be installed to a python virtual environment specific to R with:
\code{py_install("usaddress", pip = TRUE)}
This function uses a custom tag mapping to combine address components into the columns in the returned tibble
(see https://usaddress.readthedocs.io/en/latest/#details for full definition of components):
}
\details{
\itemize{
\item \code{street_number}: \code{AddressNumber}, \code{AddressNumberPrefix}, \code{AddressNumberSuffix}
\item \code{street_name}: \code{StreetName}, \code{StreetNamePreDirectional}, \code{StreetNamePostDirectional}, \code{StreetNamePostModifier}, \code{StreetNamePostType}
\item \code{city}: \code{PlaceName}
\item \code{state}: \code{StateName}
\item \code{zip}: the \strong{first five characters} of \code{ZipCode}
}

If an address is not classified as a \verb{Street Address} (i.e. \code{Intersection}, \verb{PO Box}, or \code{Ambiguous}),
then the columns in the returned component tibble will all be NA
}
